Fixed the seveal hunderd warnings in TerraPage loader!

+++ src/osgPlugins/txp/trpage_header.cpp
@@ -82,7 +82,7 @@ void trpgHeader::SetDbVersion(int32 vmaj,int32 vmin)
 }
 void trpgHeader::SetTileSize(int id,const trpg2dPoint &pt)
 {
-    if (id < 0 || id >= tileSize.size())  return;
+    if (id < 0 || static_cast<unsigned int>(id) >= tileSize.size())  return;
     tileSize[id] = pt;
 }
 void trpgHeader::SetOrigin(const trpg3dPoint &pt)
@@ -203,7 +203,7 @@ bool trpgHeader::GetDbVersion(int32 &vmaj,int32 &vmin) const
 bool trpgHeader::GetTileSize(int id,trpg2dPoint &pt) const
 {
     if (!isValid()) return false;
-    if (id < 0 || id >= tileSize.size())  return false;
+    if (id < 0 || static_cast<unsigned int>(id) >= tileSize.size())  return false;
     pt = tileSize[id];
     return true;
 }
